+++ test/amp/SortTest/SortTest.cpp
@@ -2178,11 +2178,11 @@ INSTANTIATE_TYPED_TEST_CASE_P( UDDTest, SortUDDArrayTest, UDDTests );
 
 #if (TEST_DOUBLE ==1)
 TEST (sanity_sort__withBoltClDevVectDouble_epr, floatSerial){
-	size_t sizeOfInputBufer = 64; //test case is failing for all values greater than 32
+	int sizeOfInputBufer = 64; //test case is failing for all values greater than 32
 	std::vector<double>  stdVect(0);
 	bolt::BKND::device_vector<double>  boltVect(0);
 
-	for (size_t i = 0 ; i < sizeOfInputBufer; i++){
+	for (int i = 0 ; i < sizeOfInputBufer; i++){
 	    double dValue = rand();
         dValue = dValue/rand();
         dValue = dValue*rand();
@@ -2195,6 +2195,7 @@ TEST (sanity_sort__withBoltClDevVectDouble_epr, floatSerial){
 	for (int i = 0 ; i < sizeOfInputBufer; i++){
 	    EXPECT_DOUBLE_EQ(stdVect[i], boltVect[i]);
 	}
+
 }
 #endif
 TEST (rawArrayTest, floatarray){
